MDL-29774 Fixed display of selected user's group membership information.
This is done by assigning the userid as key to each of the associated group membership information in the userSummaries array.
The issue about the user's membership info not showing when there are more than 100 potential members is resolved by populating the potentialmembersids array before checking the record count of the query results.