MDL-55324 behat: Add behat step "i_set_the_nth_field_to"
In cases where two form elements have the same label it is not
possible to target a specific one, the first selected element
will always get the value. This patch adds a new step:
"I set instance [n] of the field [x] to [y]" which will set
the nth occurance of a field matching the label x.

Corresponding:

* "Instance [n] of the field [x] matches value [y]"
* "I click on instance [n] of element [x]"
* "I click on [x] [y] in instance [n] of the [z] [w]"

Have also been added.