MDL-60940 editor: Inform the user if the forceclean flag would apply
This is to better communicate with users that the forceclean flag has
been set in the server administration. In places where there might
expect to be able to insert an interactive content (scripts, iframes
etc), we display a notification informing them that the content would be
cleaned.

Hopefully this will help to debug eventual regressions in content
displaying after setting the flag.