MDL-73679 core_courseformat: Move activity content to separate template
Course formats may want to preserve the existing activity information,
but either change the wrapper (.activity-item), or prepend and append
additional content to it.

Prior to this change, the only way to do this was by copying the entire
content of the template.

This change moves the content of the activity-item to a separate
template.

This means that a course format can replace the activity-item wrapper
whilst retaining its content by overriding the cm template and including
the cm/activity template. For example:

```
<div class="my-custom-activity-item {{!
        }}{{#modstealth}}hiddenactivity{{/modstealth}}{{!
        }}{{#modhiddenfromstudents}}hiddenactivity{{/modhiddenfromstudents}}{{!
        }}{{#modinline}}activityinline{{/modinline}}" data-activityname="{{activityname}}">
    {{#moveicon}} {{{moveicon}}} {{/moveicon}}
    {{$ core_courseformat/local/content/cm/activity }}
        {{> core_courseformat/local/content/cm/activity }}
    {{/ core_courseformat/local/content/cm/activity }}
</div>
```

It also means that a course format can append additional information to
this wrapper, for example:

```
{{< core_courseformat/local/content/cm }}
    {{$ core_courseformat/local/content/cm/activity }}
        {{> core_courseformat/local/content/cm/activity }}
    {{#cmmeta}}
    <div class="ct-activity-meta-container">
        {{{cmmeta}}}
    </div>
    {{/cmmeta}}
    {{/ core_courseformat/local/content/cm/activity }}
{{/ core_courseformat/local/content/cm }}
```